+++ src/assets/ba_data/python/bauiv1/_appsubsystem.py
@@ -4,6 +4,7 @@
 
 from __future__ import annotations
 
+import time
 import logging
 import inspect
 import weakref
@@ -84,6 +85,10 @@ class UIV1AppSubsystem(babase.AppSubsystem):
         self.title_color = (0.72, 0.7, 0.75)
         self.heading_color = (0.72, 0.7, 0.75)
         self.infotextcolor = (0.7, 0.9, 0.7)
+
+        self._last_win_recreate_size: tuple[float, float] | None = None
+        self._last_screen_size_win_recreate_time: float | None = None
+        self._screen_size_win_recreate_timer: babase.AppTimer | None = None
 
         # Elements in our root UI will call anything here when
         # activated.
@@ -394,12 +399,12 @@ class UIV1AppSubsystem(babase.AppSubsystem):
         )
 
     @override
-    def on_screen_change(self) -> None:
+    def on_ui_scale_change(self) -> None:
         # Update our stored UIScale.
         self._update_ui_scale()
 
         # Update native bits (allow root widget to rebuild itself/etc.)
-        _bauiv1.on_screen_change()
+        _bauiv1.on_ui_scale_change()
 
         # Lastly, if we have a main window, recreate it to pick up the
         # new UIScale/etc.
@@ -408,3 +413,57 @@ class UIV1AppSubsystem(babase.AppSubsystem):
             winstate = self.save_main_window_state(mainwindow)
             self.clear_main_window(transition='instant')
             self.restore_main_window_state(winstate)
+
+            # Store the size we created this for to avoid redundant
+            # future recreates.
+            self._last_win_recreate_size = babase.get_virtual_screen_size()
+
+    @override
+    def on_screen_size_change(self) -> None:
+
+        # Recreating a MainWindow is a kinda heavy thing and it doesn't
+        # seem like we should be doing it at 120hz during a live window
+        # resize, so let's limit the max rate we do it.
+        now = time.monotonic()
+
+        # 4 refreshes per second seems reasonable.
+        interval = 0.25
+
+        # If there is a timer set already, do nothing.
+        if self._screen_size_win_recreate_timer is not None:
+            return
+
+        # Ok; there's no timer. Schedule one.
+        till_update = (
+            0.0
+            if self._last_screen_size_win_recreate_time is None
+            else max(
+                0.0, self._last_screen_size_win_recreate_time + interval - now
+            )
+        )
+        self._screen_size_win_recreate_timer = babase.AppTimer(
+            till_update, self._do_screen_size_win_recreate
+        )
+
+    def _do_screen_size_win_recreate(self) -> None:
+        self._last_screen_size_win_recreate_time = time.monotonic()
+        self._screen_size_win_recreate_timer = None
+
+        # Avoid recreating if we're already at this size. This prevents
+        # a redundant recreate when ui scale changes.
+        virtual_screen_size = babase.get_virtual_screen_size()
+        if virtual_screen_size == self._last_win_recreate_size:
+            return
+
+        mainwindow = self.get_main_window()
+        if (
+            mainwindow is not None
+            and mainwindow.refreshes_on_screen_size_changes
+        ):
+            winstate = self.save_main_window_state(mainwindow)
+            self.clear_main_window(transition='instant')
+            self.restore_main_window_state(winstate)
+
+            # Store the size we created this for to avoid redundant
+            # future recreates.
+            self._last_win_recreate_size = virtual_screen_size