+++ src/ballistica/scene_v1/connection/connection_to_client.cc
@@ -444,7 +444,7 @@ void ConnectionToClient::HandleMessagePacket(
           if (!token_.empty()) {
             // Kick off a query to the master-server for this client's info.
             // FIXME: we need to add retries for this in case of failure.
-            g_base->plus()->ClientInfoQuery(
+            g_base->Plus()->ClientInfoQuery(
                 token_, our_handshake_player_spec_str_ + our_handshake_salt_,
                 peer_hash_, build_number_);
           }