+++ src/ballistica/scene_v1/python/methods/python_methods_scene.cc
@@ -1367,7 +1367,7 @@ static auto PyGetGameRoster(PyObject* self, PyObject* args, PyObject* keywds)
     // Let's also include a public account-id if we have one.
     std::string account_id;
     if (clientid == -1) {
-      account_id = g_base->plus()->GetPublicV1AccountID();
+      account_id = g_base->Plus()->GetPublicV1AccountID();
     } else {
       if (auto* appmode = classic::ClassicAppMode::GetActiveOrWarn()) {
         auto client2 =