+++ src/ballistica/scene_v1/support/scene_v1_input_device_delegate.cc
@@ -244,10 +244,10 @@ void SceneV1InputDeviceDelegate::InvalidateConnectionToHost() {
 auto SceneV1InputDeviceDelegate::GetPublicV1AccountID() const -> std::string {
   assert(g_base->InLogicThread());
 
-  // This default implementation assumes the device is local
-  // so just returns the locally signed in account's public id.
+  // This default implementation assumes the device is local so just returns
+  // the locally signed in account's public id.
 
-  return g_base->plus()->GetPublicV1AccountID();
+  return g_base->Plus()->GetPublicV1AccountID();
 }
 
 auto SceneV1InputDeviceDelegate::GetPlayerProfiles() const -> PyObject* {